moved functions to the content service
