+++ lib/api/online/challenge-io.js
@@ -23,7 +23,7 @@ module.exports = function (config) {
                 },
                 campaignRef : {
                     code        : world,
-                    challengeNo : challengeId
+                    challengeId : challengeId
                 }
             };
