+++ lib/controller/challenge.js
@@ -16,7 +16,7 @@ var app = require('../app'),
  * Controller for export modal
  */
 
-app.controller('ChallengeController', function ($scope, $routeParams, $window, $timeout, $rootScope, $location, contentService, securityService) {
+app.controller('ChallengeController', function ($scope, $routeParams, $window, $timeout, $rootScope, $location, contentService) {
     var win = angular.element($window),
         hintTimer,
         worldId;
@@ -71,7 +71,7 @@ app.controller('ChallengeController', function ($scope, $routeParams, $window, $
      * Asynchronously loads the challenge
      */
     function loadChallenge() {
-        contentService.getChallenge($scope.worldId, $scope.id).then(function (challenge) {
+        contentService.challenge.get($scope.worldId, $scope.id).then(function (challenge) {
             var lsCode;
 
             if ($scope.isChallengeLocked(challenge)) {
@@ -123,7 +123,7 @@ app.controller('ChallengeController', function ($scope, $routeParams, $window, $
 
         if (!$rootScope.selectedWorld) {
             //You're hitting the address directly, and we need to load the world's information
-            contentService.getWorld("worlds/" + $scope.worldId).then(function (data) {
+            contentService.world.get("worlds/" + $scope.worldId).then(function (data) {
                 $scope.challenges = data.challenges;
                 $rootScope.selectedWorld = data;
                 worldId = $rootScope.selectedWorld.id;
@@ -137,7 +137,7 @@ app.controller('ChallengeController', function ($scope, $routeParams, $window, $
         }
 
         if (!$rootScope.inWorldProgress) {
-            $rootScope.inWorldProgress = securityService.progress.get($scope.worldId);
+            $rootScope.inWorldProgress = contentService.progress.get($scope.worldId);
         }
     }
 
@@ -184,9 +184,7 @@ app.controller('ChallengeController', function ($scope, $routeParams, $window, $
         }
     });
 
-    $scope.isChallengeLocked = function (challenge) {
-        return securityService.challenge.isLocked(challenge);
-    };
+    $scope.isChallengeLocked = contentService.challenge.isLocked;
 
     /*
      * Get back to first step
