+++ lib/controller/challenges.js
@@ -9,7 +9,7 @@ var api,
  * Controller for challenges selection screen
  */
 
-app.controller('ChallengesController', function ($scope, $rootScope, $routeParams, $location, contentService, securityService) {
+app.controller('ChallengesController', function ($scope, $rootScope, $routeParams, $location, contentService) {
     var config = window.CONFIG,
         domain = window.location.hostname,
         domainCfg = (config.DOMAIN_CFG) ? config.DOMAIN_CFG[domain] : null,
@@ -28,7 +28,7 @@ app.controller('ChallengesController', function ($scope, $rootScope, $routeParam
      * @return void
      */
     function init() {
-        contentService.getWorlds().then(function gotWorlds(data) {
+        contentService.world.getAll().then(function gotWorlds(data) {
             $rootScope.worlds = data;
             if (routeWorldId) {
                 //a world has  been selected through the URL
@@ -63,6 +63,7 @@ app.controller('ChallengesController', function ($scope, $rootScope, $routeParam
             });
         }
     }
+
     /**
      * Loads all the information for a world
      * @param  {object} world the object that contains the world information
@@ -70,11 +71,11 @@ app.controller('ChallengesController', function ($scope, $rootScope, $routeParam
     $scope.loadWorld = function (world) {
         var worldPath = "worlds/" + world.id;
         if (!$scope.isWorldLocked(world)) {
-            contentService.getWorld(worldPath).then(function (data) {
+            contentService.world.get(worldPath).then(function (data) {
                 $scope.challenges = data.challenges;
                 $rootScope.selectedWorld = data;
                 $rootScope.selectedWorldClass = $rootScope.selectedWorld.css_class;
-                $rootScope.inWorldProgress = securityService.progress.get($rootScope.selectedWorld.id);
+                $rootScope.inWorldProgress = contentService.progress.get($rootScope.selectedWorld.id);
                 $rootScope.$broadcast('world-loaded', data);
             });
         }
@@ -85,92 +86,6 @@ app.controller('ChallengesController', function ($scope, $rootScope, $routeParam
         $rootScope.selectedWorldClass = undefined;
         $rootScope.inWorldProgress = {challengeNo: 1};
         $location.path('/challenges');
-    };
-
-    /**
-     * Returns true if the world is visible.
-     * @param  {object}  world The world
-     * @return {Boolean}       true if the world is completed
-     */
-    $scope.isWorldInIndex = function (world) {
-        return (world.visibility !== 'hidden');
-    };
-
-    /*
-     * Open Kano World browse popup if offline, proceed with normal event if online
-     *
-     * @return void
-     */
-    $scope.browseMore = function (e) {
-        if ($rootScope.cfg.OFFLINE) {
-            e.preventDefault();
-            api.challengeIO.browseMore();
-        }
-    };
-
-    /**
-     * Returns true if the world is completed.
-     * @param  {object}  world The world
-     * @return {Boolean}       true if the world is completed
-     */
-    $scope.isWorldCompleted = function (world) {
-        return securityService.world.isCompleted(world);
-    };
-
-    /**
-    *Returns true if a world is locked;
-    */
-    $scope.isWorldLocked = function (world) {
-        return securityService.world.isLocked(world);
-    };
-
-    /*
-     * Returns true if world is current.
-     * @param {object} world
-     * @return {Boolean}
-     */
-    $scope.isWorldCurrent = function (world) {
-        return securityService.world.isCurrent(world);
-    };
-
-    /*
-     * Returns true if challenge is locked
-     *
-     * @param {object} the challenge item
-     * @return {Boolean}
-     */
-    $scope.isLocked = function (challenge) {
-        return securityService.challenge.isLocked(challenge);
-    };
-
-    /*
-     * Returns true if challenge is disabled
-     *
-     * @param {Number} index
-     * @return {Boolean}
-     */
-    $scope.isCompleted = function (challenge) {
-        return securityService.challenge.isCompleted(challenge);
-    };
-
-    /*
-     * Returns true if challenge is last unlocked
-     *
-     * @param {Number} index
-     * @return {Boolean}
-     */
-    $scope.isCurrent = function (challenge) {
-        return securityService.challenge.isCurrent(challenge);
-    };
-
-    /*
-     * Returns true if all challenges have been unlocked
-     *
-     * @param {Number} index
-     * @return {Boolean}
-     */
-    $scope.isLastUnlocked = function (index) {
-        return index === inWorldProgress.challengeNo;
     };
 
     /*
@@ -199,4 +114,65 @@ app.controller('ChallengesController', function ($scope, $rootScope, $routeParam
         $scope.selectedChallenge = null;
     };
 
+    /*
+     * Open Kano World browse popup if offline, proceed with normal event if online
+     *
+     * @return void
+     */
+    $scope.browseMore = function (e) {
+        if ($rootScope.cfg.OFFLINE) {
+            e.preventDefault();
+            api.challengeIO.browseMore();
+        }
+    };
+
+    /**
+     * Returns true if the world is visible.
+     * @param  {object}  world The world
+     * @return {Boolean}       true if the world is completed
+     */
+    $scope.isWorldInIndex = contentService.world.isVisible;
+
+    /**
+     * Returns true if the world is completed.
+     * @param  {object}  world The world
+     * @return {Boolean}       true if the world is completed
+     */
+    $scope.isWorldCompleted = contentService.world.isCompleted;
+
+    /**
+    *Returns true if a world is locked;
+    */
+    $scope.isWorldLocked = contentService.world.isLocked;
+
+    /*
+     * Returns true if world is current.
+     * @param {object} world
+     * @return {Boolean}
+     */
+    $scope.isWorldCurrent = contentService.world.isCurrent;
+
+    /*
+     * Returns true if challenge is locked
+     *
+     * @param {object} the challenge item
+     * @return {Boolean}
+     */
+    $scope.isLocked = contentService.challenge.isLocked;
+
+    /*
+     * Returns true if challenge is disabled
+     *
+     * @param {Number} index
+     * @return {Boolean}
+     */
+    $scope.isCompleted = contentService.challenge.isCompleted;
+
+    /*
+     * Returns true if challenge is last unlocked
+     *
+     * @param {Number} index
+     * @return {Boolean}
+     */
+    $scope.isCurrent = contentService.challenge.isCurrent;
 });
