+++ lib/index.js
@@ -15,7 +15,6 @@ require('./routes');
 // Service
 require('./service/email');
 require('./service/content');
-require('./service/security');
 
 // Controllers
 require('./controller/main');
